docs(README): redirect to upstream previews
I think it's fine for the previews to be
redirected upstream since it's a repo under
our own organisation.

In the future, I might extend my GitHub action
to support automatic replacement of this tag but
I don't think it's too much of a concern if this
preview becomes out of date every now and then.